# Conflicts:
#	plugins/org.jkiss.dbeaver.ext.postgresql.ui/src/org/jkiss/dbeaver/ext/postgresql/ui/config/PostgreProcedureConfigurator.java
#	plugins/org.jkiss.dbeaver.ext.postgresql/src/org/jkiss/dbeaver/ext/postgresql/PostgreUtils.java
#	plugins/org.jkiss.dbeaver.ext.postgresql/src/org/jkiss/dbeaver/ext/postgresql/edit/PostgreCommandGrantPrivilege.java
#	plugins/org.jkiss.dbeaver.ext.postgresql/src/org/jkiss/dbeaver/ext/postgresql/edit/PostgreForeignKeyManager.java
#	plugins/org.jkiss.dbeaver.ext.postgresql/src/org/jkiss/dbeaver/ext/postgresql/edit/PostgreTableManager.java
#	plugins/org.jkiss.dbeaver.ext.postgresql/src/org/jkiss/dbeaver/ext/postgresql/edit/PostgreTableManagerBase.java
#	plugins/org.jkiss.dbeaver.ext.postgresql/src/org/jkiss/dbeaver/ext/postgresql/model/PostgreDialect.java
#	plugins/org.jkiss.dbeaver.ext.postgresql/src/org/jkiss/dbeaver/ext/postgresql/model/PostgreProcedure.java
#	plugins/org.jkiss.dbeaver.ext.postgresql/src/org/jkiss/dbeaver/ext/postgresql/model/PostgreSequence.java
#	plugins/org.jkiss.dbeaver.model/src/org/jkiss/dbeaver/model/struct/DBStructUtils.java
#	plugins/org.jkiss.dbeaver.registry/src/org/jkiss/dbeaver/registry/DataSourceDescriptor.java
#	plugins/org.jkiss.dbeaver.ui.app.standalone/splash.png
#	plugins/org.jkiss.dbeaver.ui.editors.data/src/org/jkiss/dbeaver/ui/controls/resultset/spreadsheet/Spreadsheet.java
#	product/debug/DBeaver.product.launch

+++ plugins/org.jkiss.dbeaver.core/src/org/jkiss/dbeaver/tools/project/ProjectExportWizard.java
@@ -1,6 +1,6 @@
 /*
  * DBeaver - Universal Database Manager
- * Copyright (C) 2010-2023 DBeaver Corp and others
+ * Copyright (C) 2010-2024 DBeaver Corp and others
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -119,6 +119,19 @@ public class ProjectExportWizard extends Wizard implements IExportWizard {
 
         String archiveName = exportData.getArchiveFileName() + ExportConstants.ARCHIVE_FILE_EXT;
         File archiveFile = new File(exportData.getOutputFolder(), archiveName);
+        
+        if (archiveFile.exists()) {
+            boolean overwrite = DBWorkbench.getPlatformUI().confirmAction(
+                CoreMessages.dialog_project_export_wizard_file_overwrite_window_title,
+                NLS.bind(CoreMessages.dialog_project_export_wizard_file_overwrite_confirm, archiveName),
+                true
+            );
+
+            if (!overwrite) {
+                return;
+            }
+        }
+        
         FileOutputStream exportStream = new FileOutputStream(archiveFile);
 
         try {
@@ -134,10 +147,11 @@ public class ProjectExportWizard extends Wizard implements IExportWizard {
 
             {
                 // Export source info
+                InetAddress localHost = RuntimeUtils.getLocalHostOrLoopback();
                 meta.startElement(ExportConstants.TAG_SOURCE);
                 meta.addAttribute(ExportConstants.ATTR_TIME, System.currentTimeMillis());
-                meta.addAttribute(ExportConstants.ATTR_ADDRESS, InetAddress.getLocalHost().getHostAddress());
-                meta.addAttribute(ExportConstants.ATTR_HOST, InetAddress.getLocalHost().getHostName());
+                meta.addAttribute(ExportConstants.ATTR_ADDRESS, localHost.getHostAddress());
+                meta.addAttribute(ExportConstants.ATTR_HOST, localHost.getHostName());
                 meta.endElement();
             }
